Refactored folder structure for views
