+++ src/Components/App/AddPage.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useContext } from "react";
 import { v4 as uuidv4 } from "uuid";
 
-import Button from "../Views/Button";
-import Chip from "../Views/Chip";
-import { AddIcon, CloseIcon } from "../Views/Icons";
+import Button from "../../views/Button";
+import Chip from "../../views/Chip";
+import { AddIcon, CloseIcon } from "../../views/Icons";
 
 import { FirebaseContext } from "../Firebase";
 import { motion } from "framer-motion";
