+++ src/Components/App/SignIn.jsx
@@ -1,8 +1,8 @@
-import Logo from "../Views/Logo";
-import Button from "../Views/Button";
+import Logo from "../../views/Logo";
+import Button from "../../views/Button";
 import { useContext } from "react";
 import { FirebaseContext } from "../Firebase";
-import { GoogleIcon } from "../Views/Icons";
+import { GoogleIcon } from "../../views/Icons";
 
 const SignInPage = () => {
   const { signInGoogle } = useContext(FirebaseContext);
